refactor: have a single module called ethereum
