+++ crypto/usdc-ethereum/cmd/main.go
@@ -2,16 +2,15 @@ package main
 
 import (
 	"fmt"
+	"github.com/julianespinel/lab/crypto/usdc-ethereum/internal/ethereum/clients"
+	"github.com/julianespinel/lab/crypto/usdc-ethereum/internal/ethereum/models"
 	"log"
 
 	"github.com/julianespinel/lab/crypto/usdc-ethereum/config"
-	"github.com/julianespinel/lab/crypto/usdc-ethereum/internal"
-	"github.com/julianespinel/lab/crypto/usdc-ethereum/internal/clients"
-	"github.com/julianespinel/lab/crypto/usdc-ethereum/internal/models"
-	"github.com/julianespinel/lab/crypto/usdc-ethereum/internal/usdc"
+	"github.com/julianespinel/lab/crypto/usdc-ethereum/internal/ethereum"
 )
 
-func fetchAndProcessContractEvents(usdcService *usdc.USDCService, cfg *config.Config) ([]models.EventLog, error) {
+func fetchAndProcessContractEvents(usdcService *ethereum.USDCService, cfg *config.Config) ([]models.EventLog, error) {
 	fmt.Println("Fetching USDC events between:", cfg.StartDate, "and", cfg.EndDate)
 	events, err := usdcService.FetchUSDCContractEventsByDateRange(cfg.StartDate, cfg.EndDate, cfg.USDCContractAddress)
 	if err != nil {
@@ -22,14 +21,14 @@ func fetchAndProcessContractEvents(usdcService *usdc.USDCService, cfg *config.Co
 		return nil, fmt.Errorf("no USDC events found")
 	}
 
-	internal.DisplayEvents(events)
-	internal.CalculateTotalAmount(events)
-	internal.CalculateTotalAmountPerAddress(events)
+	ethereum.DisplayEvents(events)
+	ethereum.CalculateTotalAmount(events)
+	ethereum.CalculateTotalAmountPerAddress(events)
 
 	return events, nil
 }
 
-func fetchAndProcessWalletTransactions(usdcService *usdc.USDCService, cfg *config.Config, numTransactions int) ([]models.EventLog, error) {
+func fetchAndProcessWalletTransactions(usdcService *ethereum.USDCService, cfg *config.Config, numTransactions int) ([]models.EventLog, error) {
 	fmt.Println("Fetching transactions for wallet:", cfg.WalletAddress)
 
 	events, err := usdcService.FetchLastTransactionsFromWallet(cfg.WalletAddress, numTransactions)
@@ -41,9 +40,9 @@ func fetchAndProcessWalletTransactions(usdcService *usdc.USDCService, cfg *confi
 		return nil, fmt.Errorf("no wallet events found")
 	}
 
-	internal.DisplayEvents(events)
-	internal.CalculateTotalAmount(events)
-	internal.CalculateTotalAmountPerAddress(events)
+	ethereum.DisplayEvents(events)
+	ethereum.CalculateTotalAmount(events)
+	ethereum.CalculateTotalAmountPerAddress(events)
 
 	return events, nil
 }
@@ -63,7 +62,7 @@ func main() {
 	defer ethClient.Close()
 
 	// Initialize USDC service
-	usdcService := usdc.NewUSDCService(ethClient, cfg.EtherscanAPIKey)
+	usdcService := ethereum.NewUSDCService(ethClient, cfg.EtherscanAPIKey)
 
 	_, err = fetchAndProcessContractEvents(usdcService, cfg)
 	if err != nil {
