+++ crypto/usdc-ethereum/internal/ethereum/models/event.go
